WhatsApp-Bot/requirements.txt
flask
twilio
requests
googlesearch-python
